Display error if basic_auth worked.
We run `@client.authorizations` to list the clients authorizations.
However, it's pointless because we throw away the output.

We just run this to see if it works. If it does, then the client is
logged in; otherwise, the login failed.

TODO: Find a better way to do this, w/o making pointless requests.