Finally got it to commit!
I was basing my commit off of this guide:
http://mattgreensmith.net/2013/08/08/commit-directly-to-github-via-api-with-octokit/

Problem was, the guide was based off of the v2 API. The current API is v3!
The function create_tree doesn't need the base_tree that I was giving
it. That's why I kept on getting a "422 - Invalid tree info" error.

Although, it still doesn't work fully. For some reason it deletes files
that aren't changed. I'll try to fix that in the next commit.